Remove deprecated code related to chat export
